several things. see long git message
replaced usage of 3.141.... with Math.PI
replaced usage of 111.2 with 111.76 which I believe is a more accurate
value for the length in km for a degree of latitude

created function "longitudeDistAtLatitude" to get the width in km of 1 degree of
latitude. needs the corresponding latitude because of how balls work
note: function hardcodes latitude as 55 because of unknown issue

modified function "latlongToCartesianXZ" to use the negative of
longitude.
previously a more positive patitude meant the tile went further
in positive X.
now a more positive latitude means it goes further in NEGATIVE X.
this may seem counter-intuitive, which is why it was like that to start
with, but it is needed in order to maintain the property that anincrease
in LONGITUDE to coresponds to an increase in Z position.

fixed my interpretation of lat/long values from bounding box variable
(the latitude was previously taken from index 0, but it actually resides
in index 1

accounted for the fact that threejs plane origin is in the center,
accounted for by adding to the latitude and longitude half of the
respective difference from corner to corner.

swapped the width and height for the threejs plane, and ensured it
gets rotated -PI/2 on X & +PI on z.
this is just to position the plane's north edge to face positive Z

once again adusted camera starting position to lat,long 55,-4 so it's
less annoying to get to the relevant tiles.

created "gimmeButton" which just contains abstracted button creation code,
so far needed only for development purposes

added utility button which displays the current camera position,
and permits setting camera position via dialogue box when clicked.

wrote and then commented out code which rotates the mountain data array
some wierd way. hopefully wont be needed but not deleting just yet.
^ it got commented because it was part of a fruitless attempt at a fix